requirements.txt
streamlit==0.80.0
pandas==1.1.5
numpy==1.19.5
Pillow==8.2.0
